+++ Source/CesiumRuntime/Public/Cesium3DTileset.h
@@ -20,6 +20,7 @@
 
 class UMaterialInterface;
 class ACesiumCartographicSelection;
+struct FCesiumCamera;
 
 namespace Cesium3DTilesSelection {
 class Tileset;
@@ -342,8 +343,12 @@ public:
    *
    * This is an experimental feature and may change in future versions.
    */
-  UPROPERTY(EditAnywhere, Category = "Cesium|Experimental")
-  TArray<FCesiumExclusionZone> ExclusionZones;
+  UPROPERTY(
+      meta =
+          (DeprecatedProperty,
+           DeprecationMessage =
+               "Exclusion Zones have been deprecated. Please use Cartographic Polygon actor instead."))
+  TArray<FCesiumExclusionZone> ExclusionZones_DEPRECATED;
 
   /**
    * The screen-space error to be enforced for tiles that are outside the view
@@ -460,6 +465,19 @@ private:
       Category = "Cesium",
       meta = (EditCondition = "TilesetSource==ETilesetSource::FromCesiumIon"))
   FString IonAccessToken;
+
+  /**
+   * The URL of the ion asset endpoint. Defaults to Cesium ion but a custom
+   * endpoint can be specified.
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintGetter = GetIonAssetEndpointUrl,
+      BlueprintSetter = SetIonAssetEndpointUrl,
+      Category = "Cesium",
+      AdvancedDisplay,
+      meta = (EditCondition = "TilesetSource==ETilesetSource::FromCesiumIon"))
+  FString IonAssetEndpointUrl;
 
   /**
    * Check if the Cesium ion token used to access this tileset is working
@@ -589,7 +607,7 @@ public:
   FString GetUrl() const { return Url; }
 
   UFUNCTION(BlueprintSetter, Category = "Cesium")
-  void SetUrl(FString InUrl);
+  void SetUrl(const FString& InUrl);
 
   UFUNCTION(BlueprintGetter, Category = "Cesium")
   int64 GetIonAssetID() const { return IonAssetID; }
@@ -601,7 +619,13 @@ public:
   FString GetIonAccessToken() const { return IonAccessToken; }
 
   UFUNCTION(BlueprintSetter, Category = "Cesium")
-  void SetIonAccessToken(FString InAccessToken);
+  void SetIonAccessToken(const FString& InAccessToken);
+
+  UFUNCTION(BlueprintGetter, Category = "Cesium")
+  FString GetIonAssetEndpointUrl() const { return IonAssetEndpointUrl; }
+
+  UFUNCTION(BlueprintSetter, Category = "Cesium")
+  void SetIonAssetEndpointUrl(const FString& InIonAssetEndpointUrl);
 
   UFUNCTION(BlueprintGetter, Category = "Cesium|Physics")
   bool GetCreatePhysicsMeshes() const { return CreatePhysicsMeshes; }
@@ -712,25 +736,13 @@ private:
   void LoadTileset();
   void DestroyTileset();
 
-  struct UnrealCameraParameters {
-    FVector2D viewportSize;
-    FVector location;
-    FRotator rotation;
-    double fieldOfViewDegrees;
-
-    // If not std::nullopt, the aspect ratio may be different from the one
-    // implied by the viewportSize and black bars are added as needed in order
-    // to achieve this aspect ratio within a larger viewport.
-    std::optional<double> aspectRatio = std::nullopt;
-  };
-
   static Cesium3DTilesSelection::ViewState CreateViewStateFromViewParameters(
-      const UnrealCameraParameters& camera,
+      const FCesiumCamera& camera,
       const glm::dmat4& unrealWorldToTileset);
 
-  std::vector<UnrealCameraParameters> GetCameras() const;
-  std::vector<UnrealCameraParameters> GetPlayerCameras() const;
-  std::vector<UnrealCameraParameters> GetSceneCaptures() const;
+  std::vector<FCesiumCamera> GetCameras() const;
+  std::vector<FCesiumCamera> GetPlayerCameras() const;
+  std::vector<FCesiumCamera> GetSceneCaptures() const;
 
 public:
   /**
@@ -777,7 +789,7 @@ private:
   void AddFocusViewportDelegate();
 
 #if WITH_EDITOR
-  std::vector<UnrealCameraParameters> GetEditorCameras() const;
+  std::vector<FCesiumCamera> GetEditorCameras() const;
 
   /**
    * Will focus all viewports on this tileset.